+++ .travis.yml
@@ -1,4 +1,3 @@
-sudo: false
 language: scala
 
 scala:
@@ -21,13 +20,4 @@ script:
 
 after_success:
   - sbt coveralls
-
-cache:
-  directories:
-    - $HOME/.ivy2/cache
-    - $HOME/.sbt/boot/
-
-before_cache:
-  - find $HOME/.ivy2 -name "ivydata-*.properties" -delete
-  - find $HOME/.sbt -name "*.lock" -delete
