+++ docs/src/site/_config.yml
@@ -1,0 +1,25 @@
+# Setup
+title:        finagle-postgres
+tagline:      Asynchronous Scala PostgreSQL client using Finagle
+baseurl:      "/finagle-postgres"
+#paginate:     1
+
+# About/contact
+author:
+  name:       Jeremy Smith
+  url:        http://jeremyrsmith.github.io 
+
+# Gems
+gems:
+  - jekyll-paginate
+  - kramdown
+  - rouge
+
+#Others
+markdown: kramdown
+
+kramdown:
+  input: GFM
+  syntax_highlighter: rouge
+  highlighter: rouge
+
