+++ docs/src/site/_includes/page-header.html
@@ -1,0 +1,5 @@
+<section class="page-header">
+  <h1 class="project-name">{{ site.title }}</h1>
+  <h2 class="project-tagline">{{ site.tagline }}</h2>
+  <a href="https://github.com/jeremyrsmith/featherbed" class="btn">View on GitHub</a>
+</section>
