+++ docs/src/site/_layouts/post.html
@@ -1,0 +1,7 @@
+---
+layout: default
+---
+<h2>{{ page.title }}</h2>
+<p class="meta">{{ page.date | date_to_string }}</p>
+
+{{ content }}
