+++ docs/src/site/css/code.css
@@ -1,0 +1,61 @@
+.main-content pre.highlight code .hll { background-color: #ffffcc }
+.main-content pre.highlight code .c { color: #999988; font-style: italic } /* Comment */
+.main-content pre.highlight code .err { color: #a61717; background-color: #e3d2d2 } /* Error */
+.main-content pre.highlight code .k { color: #000000; font-weight: bold } /* Keyword */
+.main-content pre.highlight code .o { color: #000000; font-weight: bold } /* Operator */
+.main-content pre.highlight code .cm { color: #999988; font-style: italic } /* Comment.Multiline */
+.main-content pre.highlight code .cp { color: #999999; font-weight: bold; font-style: italic } /* Comment.Preproc */
+.main-content pre.highlight code .c1 { color: #999988; font-style: italic } /* Comment.Single */
+.main-content pre.highlight code .cs { color: #999999; font-weight: bold; font-style: italic } /* Comment.Special */
+.main-content pre.highlight code .gd { color: #000000; background-color: #ffdddd } /* Generic.Deleted */
+.main-content pre.highlight code .ge { color: #000000; font-style: italic } /* Generic.Emph */
+.main-content pre.highlight code .gr { color: #aa0000 } /* Generic.Error */
+.main-content pre.highlight code .gh { color: #999999 } /* Generic.Heading */
+.main-content pre.highlight code .gi { color: #000000; background-color: #ddffdd } /* Generic.Inserted */
+.main-content pre.highlight code .go { color: #888888 } /* Generic.Output */
+.main-content pre.highlight code .gp { color: #555555 } /* Generic.Prompt */
+.main-content pre.highlight code .gs { font-weight: bold } /* Generic.Strong */
+.main-content pre.highlight code .gu { color: #aaaaaa } /* Generic.Subheading */
+.main-content pre.highlight code .gt { color: #aa0000 } /* Generic.Traceback */
+.main-content pre.highlight code .kc { color: #000000; font-weight: bold } /* Keyword.Constant */
+.main-content pre.highlight code .kd { color: #000000; font-weight: bold } /* Keyword.Declaration */
+.main-content pre.highlight code .kn { color: #000000; font-weight: bold } /* Keyword.Namespace */
+.main-content pre.highlight code .kp { color: #000000; font-weight: bold } /* Keyword.Pseudo */
+.main-content pre.highlight code .kr { color: #000000; font-weight: bold } /* Keyword.Reserved */
+.main-content pre.highlight code .kt { color: #445588; font-weight: bold } /* Keyword.Type */
+.main-content pre.highlight code .m { color: #009999 } /* Literal.Number */
+.main-content pre.highlight code .s { color: #d01040 } /* Literal.String */
+.main-content pre.highlight code .na { color: #008080 } /* Name.Attribute */
+.main-content pre.highlight code .nb { color: #0086B3 } /* Name.Builtin */
+.main-content pre.highlight code .nc { color: #445588; font-weight: bold } /* Name.Class */
+.main-content pre.highlight code .no { color: #008080 } /* Name.Constant */
+.main-content pre.highlight code .nd { color: #3c5d5d; font-weight: bold } /* Name.Decorator */
+.main-content pre.highlight code .ni { color: #800080 } /* Name.Entity */
+.main-content pre.highlight code .ne { color: #990000; font-weight: bold } /* Name.Exception */
+.main-content pre.highlight code .nf { color: #990000; font-weight: bold } /* Name.Function */
+.main-content pre.highlight code .nl { color: #990000; font-weight: bold } /* Name.Label */
+.main-content pre.highlight code .nn { color: #555555 } /* Name.Namespace */
+.main-content pre.highlight code .nt { color: #000080 } /* Name.Tag */
+.main-content pre.highlight code .nv { color: #008080 } /* Name.Variable */
+.main-content pre.highlight code .ow { color: #000000; font-weight: bold } /* Operator.Word */
+.main-content pre.highlight code .w { color: #bbbbbb } /* Text.Whitespace */
+.main-content pre.highlight code .mf { color: #009999 } /* Literal.Number.Float */
+.main-content pre.highlight code .mh { color: #009999 } /* Literal.Number.Hex */
+.main-content pre.highlight code .mi { color: #009999 } /* Literal.Number.Integer */
+.main-content pre.highlight code .mo { color: #009999 } /* Literal.Number.Oct */
+.main-content pre.highlight code .sb { color: #d01040 } /* Literal.String.Backtick */
+.main-content pre.highlight code .sc { color: #d01040 } /* Literal.String.Char */
+.main-content pre.highlight code .sd { color: #d01040 } /* Literal.String.Doc */
+.main-content pre.highlight code .s2 { color: #d01040 } /* Literal.String.Double */
+.main-content pre.highlight code .se { color: #d01040 } /* Literal.String.Escape */
+.main-content pre.highlight code .sh { color: #d01040 } /* Literal.String.Heredoc */
+.main-content pre.highlight code .si { color: #d01040 } /* Literal.String.Interpol */
+.main-content pre.highlight code .sx { color: #d01040 } /* Literal.String.Other */
+.main-content pre.highlight code .sr { color: #009926 } /* Literal.String.Regex */
+.main-content pre.highlight code .s1 { color: #d01040 } /* Literal.String.Single */
+.main-content pre.highlight code .ss { color: #990073 } /* Literal.String.Symbol */
+.main-content pre.highlight code .bp { color: #999999 } /* Name.Builtin.Pseudo */
+.main-content pre.highlight code .vc { color: #008080 } /* Name.Variable.Class */
+.main-content pre.highlight code .vg { color: #008080 } /* Name.Variable.Global */
+.main-content pre.highlight code .vi { color: #008080 } /* Name.Variable.Instance */
+.main-content pre.highlight code .il { color: #009999 } /* Literal.Number.Integer.Long */
