+++ project/plugins.sbt
@@ -1,9 +1,14 @@
 resolvers ++= Seq(
   Classpaths.typesafeReleases,
-  Classpaths.sbtPluginReleases,
-  "jgit-repo" at "http://download.eclipse.org/jgit/maven"
+  Classpaths.sbtPluginReleases
 )
 
+addSbtPlugin("org.tpolecat" % "tut-plugin" % "0.4.6")
+addSbtPlugin("org.xerial.sbt" % "sbt-sonatype" % "1.1")
 addSbtPlugin("com.jsuereth" % "sbt-pgp" % "1.0.0")
 addSbtPlugin("org.scoverage" % "sbt-scoverage" % "1.0.4")
 addSbtPlugin("org.scoverage" % "sbt-coveralls" % "1.0.0.BETA1")
+
+addSbtPlugin("com.typesafe.sbt" % "sbt-ghpages" % "0.5.4")
+addSbtPlugin("com.typesafe.sbt" % "sbt-site" % "1.1.0")
+addSbtPlugin("com.eed3si9n" % "sbt-unidoc" % "0.3.3")
